+++ Packages/Accelib/Module/Accelib.UI.Popup/Runtime/Utility/PopupOpener_Modal.cs
@@ -1,15 +1,18 @@
-﻿using Accelib.Module.UI.Popup.Data;
-using Accelib.Module.UI.Popup.Layer;
+﻿using Accelib.UI.Popup.Runtime.Data;
+using Accelib.UI.Popup.Runtime.Layer.Base;
 using Sirenix.OdinInspector;
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace Accelib.Module.UI.Popup.Utility
+namespace Accelib.UI.Popup.Runtime.Utility
 {
+    /// <summary>
+    /// UnityEvent 기반 모달 열기 헬퍼.
+    /// </summary>
     public class PopupOpener_Modal : MonoBehaviour
     {
         [Header("Text(Or Key)")]
-        [SerializeField] private ModalOpenOption option;
+        [SerializeReference] private SO_ModalOpenOption option;
 
         [Header("Events")]
         public UnityEvent onOK;